Misc ADB over WiFi
Fixup code. adbd doesnt need to be restarted when setting the port, setting to 5555 allows connections, -1 disables.

Add current WiFi IP to ADB over WiFi description text to make it easier to find.

Move network info to the top of the misc section (for ease of seeing the IP)

Conflicts:
	app/src/main/java/com/grarak/kerneladiutor/fragments/kernel/MiscFragment.java
	app/src/main/res/values/strings.xml