Screen: Initial Color Control Profiles
These three seem to be working as intended. Most of this right now's a big hack to get the concept working.

Will work on this more later.

Currently, RGB Values are good to pass.

Most of the contrast, saturation and value values need to be modified as follows:
contrast + 128, value + 128, sat + 225
from what they would typically be posted as.